# src/schemas/openapi.yaml
openapi: 3.0.3
info:
  title: Discord Last.fm Now Playing Bot - 踏み台サーバーAPI
  description: |
    Last.fm APIを使用してDiscordのリッチプレゼンスに現在再生中の楽曲を表示し、
    Discord Botでチャンネル通知と統計レポートを送信するアプリケーションの踏み台サーバーAPI。
    
    ## 機能
    - 現在再生中の楽曲情報の取得
    - 直近の再生履歴の取得（件数指定・期間指定・ページネーション対応）
    - 音楽レポートの取得（日次・週次・月次）
    - WebSocketによるリアルタイム通信
    - ヘルスチェック機能
    - レート制限機能（1分間100リクエスト）
    
    ## WebSocket
    WebSocket接続: `ws://localhost:3001`
    
    リアルタイムで以下の情報を受信できます：
    - 楽曲変更通知
    - レポート生成状況
    - 接続ステータス
  version: 1.0.0
  contact:
    name: Discord Last.fm Now Playing Bot
    url: https://github.com/kakeru-ikeda/nowplaying-for-discord
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: ローカル開発サーバー
paths:
...
  /api/now-playing:
    get:
      summary: 現在再生中の楽曲情報取得
      description: Last.fm APIから現在再生中の楽曲情報を取得
      operationId: getNowPlaying
      responses:
        '200':
          description: 楽曲情報の取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NowPlayingApiResponse'
              examples:
                playing:
                  summary: 楽曲再生中
                  value:
                    success: true
                    data:
                      artist: "スリーズブーケ"
                      track: "恋景色昭和ロマン"
                      album: "アルバム名"
                      imageUrl: "https://lastfm.freetls.fastly.net/i/u/300x300/example.png"
                      isPlaying: true
                    timestamp: "2025-07-06T10:30:00.000Z"
                not_playing:
                  summary: 楽曲停止中
                  value:
                    success: true
                    data:
                      artist: ""
                      track: ""
                      isPlaying: false
                    timestamp: "2025-07-06T10:30:00.000Z"
        '500':
          description: Last.fm API取得エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  # 日次レポート
...
  /api/recent-tracks:
    get:
      summary: 直近の再生履歴取得
      description: |
        Last.fm APIから指定した件数の直近の再生履歴を取得。
        現在再生中の楽曲も含まれ、期間指定やページネーションに対応。
      operationId: getRecentTracks
      parameters:
        - name: limit
          in: query
          description: 取得件数（1-200、デフォルト50）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: page
          in: query
          description: ページ番号（デフォルト1）
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: from
          in: query
          description: 開始日時（ISO 8601形式：YYYY-MM-DDTHH:mm:ss.sssZ）
          required: false
          schema:
            type: string
            format: date-time
          example: "2025-07-01T00:00:00.000Z"
        - name: to
          in: query
          description: 終了日時（ISO 8601形式：YYYY-MM-DDTHH:mm:ss.sssZ）
          required: false
          schema:
            type: string
            format: date-time
          example: "2025-07-06T23:59:59.999Z"
      responses:
        '200':
          description: 再生履歴の取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentTracksApiResponse'
              examples:
                recent_tracks:
                  summary: 直近の再生履歴
                  value:
                    success: true
                    data:
                      tracks:
                        - artist: "スリーズブーケ"
                          track: "恋景色昭和ロマン"
                          album: "アルバム名"
                          imageUrl: "https://lastfm.freetls.fastly.net/i/u/300x300/example.png"
                          isPlaying: true
                          url: "https://www.last.fm/music/..."
                        - artist: "Artist Name"
                          track: "Previous Track"
                          album: "Album Name"
                          imageUrl: "https://lastfm.freetls.fastly.net/i/u/300x300/example2.png"
                          isPlaying: false
                          playedAt: "2025-07-06T10:25:00.000Z"
                          url: "https://www.last.fm/music/..."
                      pagination:
                        page: 1
                        limit: 50
                        total: 2
                      period:
                        from: "2025-07-01T00:00:00.000Z"
                        to: "2025-07-06T23:59:59.999Z"
                    timestamp: "2025-07-06T10:30:00.000Z"
        '400':
          description: 不正なリクエストパラメータ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                invalid_limit:
                  summary: 不正なlimitパラメータ
                  value:
                    success: false
                    error: "Invalid limit. Must be between 1 and 200"
                    code: "INVALID_REQUEST"
                    timestamp: "2025-07-06T10:30:00.000Z"
                invalid_date:
                  summary: 不正な日付形式
                  value:
                    success: false
                    error: "Invalid from date format. Use ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ)"
                    code: "INVALID_REQUEST"
                    timestamp: "2025-07-06T10:30:00.000Z"
        '429':
          description: レート制限エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                success: false
                error: "Rate limit exceeded"
                code: "RATE_LIMIT_EXCEEDED"
                timestamp: "2025-07-06T10:30:00.000Z"
        '500':
          description: Last.fm API取得エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  # ユーザー統計情報
  /api/user-stats:
    get:
      summary: ユーザー統計情報取得
      description: Last.fm APIからユーザーの基本情報とNo.1アーティスト・トラック情報を取得
      operationId: getUserStats
      responses:
        '200':
          description: ユーザー統計情報の取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatsApiResponse'
              example:
                success: true
                data:
                  profile:
                    username: "username"
                    realName: "Real Name"
                    url: "https://www.last.fm/user/username"
                    country: "Japan"
                    registeredDate: "2015-01-01 00:00"
                    totalPlayCount: 12345
                    profileImage: "https://lastfm.freetls.fastly.net/i/u/300x300/example.png"
                  topArtist:
                    name: "アーティスト名"
                    playCount: 500
                    url: "https://www.last.fm/music/artist"
                    image: "https://lastfm.freetls.fastly.net/i/u/300x300/artist.png"
                  topTrack:
                    name: "楽曲名"
                    artist: "アーティスト名"
                    playCount: 100
                    url: "https://www.last.fm/music/track"
                    image: "https://lastfm.freetls.fastly.net/i/u/300x300/track.png"
                  generatedAt: "2025-07-06T10:30:00.000Z"
                timestamp: "2025-07-06T10:30:00.000Z"
        '500':
          description: ユーザー統計情報取得エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Failed to fetch user statistics"
                timestamp: "2025-07-06T10:30:00.000Z"
components:
  schemas:
...
    BaseApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: APIの実行結果
        timestamp:
          type: string
          format: date-time
          description: レスポンス生成時刻（ISO 8601形式）
      required:
        - success
        - timestamp

    # ヘルスチェック
...
    NowPlayingInfo:
      type: object
      properties:
        artist:
          type: string
          description: アーティスト名
        track:
          type: string
          description: 楽曲名
        album:
          type: string
          description: アルバム名
        imageUrl:
          type: string
          format: uri
          description: アルバムアートワークURL
        isPlaying:
          type: boolean
          description: 再生状態
      required:
        - artist
        - track
        - isPlaying

    # ナウプレイングAPIレスポンス
    NowPlayingApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/NowPlayingInfo'

    # 聴取推移データ
...
    ApiErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: string
              description: エラーメッセージ
            code:
              type: string
              description: エラーコード
            details:
              type: object
              description: エラー詳細情報
          required:
            - error

    # WebSocketメッセージ
...
    RecentTrackInfo:
      type: object
      properties:
        artist:
          type: string
          description: アーティスト名
        track:
          type: string
          description: 楽曲名
        album:
          type: string
          description: アルバム名
        imageUrl:
          type: string
          format: uri
          description: アルバムアートワークURL
        isPlaying:
          type: boolean
          description: 現在再生中かどうか
        playedAt:
          type: string
          format: date-time
          description: 再生日時（現在再生中の場合はnull）
        url:
          type: string
          format: uri
          description: Last.fm楽曲ページURL
      required:
        - artist
        - track
        - isPlaying

    # 再生履歴レスポンス
    RecentTracksApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                tracks:
                  type: array
                  items:
                    $ref: '#/components/schemas/RecentTrackInfo'
                  description: 再生履歴の配列
                pagination:
                  type: object
                  properties:
                    page:
                      type: integer
                      description: 現在のページ番号
                    limit:
                      type: integer
                      description: 1ページあたりの件数
                    total:
                      type: integer
                      description: 実際に取得された件数
                  required:
                    - page
                    - limit
                    - total
                period:
                  type: object
                  properties:
                    from:
                      type: string
                      format: date-time
                      description: 期間開始日時
                    to:
                      type: string
                      format: date-time
                      description: 期間終了日時
              required:
                - tracks
                - pagination

    # ユーザー統計情報
    UserStats:
      type: object
      properties:
        profile:
          type: object
          properties:
            username:
              type: string
              description: ユーザー名
            realName:
              type: string
              description: 本名（省略可能）
            url:
              type: string
              description: Last.fmプロフィールURL
            country:
              type: string
              description: 国（省略可能）
            registeredDate:
              type: string
              description: 登録日時
            totalPlayCount:
              type: number
              description: 総再生回数
            profileImage:
              type: string
              description: プロフィール画像URL（省略可能）
          required:
            - username
            - url
            - registeredDate
            - totalPlayCount
        topArtist:
          type: object
          nullable: true
          properties:
            name:
              type: string
              description: アーティスト名
            playCount:
              type: number
              description: 再生回数
            url:
              type: string
              description: アーティストページURL
            image:
              type: string
              description: アーティスト画像URL（省略可能）
          required:
            - name
            - playCount
            - url
        topTrack:
          type: object
          nullable: true
          properties:
            name:
              type: string
              description: 楽曲名
            artist:
              type: string
              description: アーティスト名
            playCount:
              type: number
              description: 再生回数
            url:
              type: string
              description: 楽曲ページURL
            image:
              type: string
              description: 楽曲画像URL（省略可能）
          required:
            - name
            - artist
            - playCount
            - url
        generatedAt:
          type: string
          format: date-time
          description: 統計情報生成時刻（ISO 8601形式）
      required:
        - profile
        - topArtist
        - topTrack
        - generatedAt
    UserStatsApiResponse:
      allOf:
        - $ref: '#/components/schemas/BaseApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserStats'

    # 楽曲情報